Fix json payload mapping
In the previous version the output of koyeb service logs was raw, previous change introduced the use of the -o json to get the stream, but this had effect of embedding the full json payload.

This change is using timestamp and a few other values from the koyeb logs, and setting the message with only the message an no longer the wrapped output from the koyeb cli json output.